Revert "Merge pull request #21 from 76696265636f646572/localPlayback-composable"
This reverts commit 9d7fe2a8213786821e5ced291aa56fdd2d12094c, reversing
changes made to 4af8185fbad1ebf6f9004a57c1089717a673a09d.
